requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
pydantic==2.11.7
python-dotenv==1.1.1
openai==1.101.0
psycopg[binary]==3.2.10
SQLAlchemy==2.0.43
feedparser==6.0.11
beautifulsoup4==4.13.4
python-dateutil==2.9.0.post0
pgvector==0.3.6
Pillow==10.4.0
trafilatura==1.12.2
alembic==1.14.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
scikit-learn==1.5.2
numpy==2.1.1
minio==7.2.9
python-multipart==0.0.20
playwright==1.55.0
